stream/stream_smb.c: fix #5936 using new api
Thread-unsafe libsmbclient crashes when used across threads, for example
when provided more than one "smb://" URI. Switch from the compatibility
api to the new libsmbclient api. Each use - presumably per-thread - of
`open_f` creates a new smb context, the deallocation of which is tied to
the talloc destructor of the stream. As opposed to locking cross-thread
calls, the new thread-safe api provides better streaming performance.
When the following thread-safety issue is fixed, this should be the
preferred solution:

* https://bugzilla.samba.org/show_bug.cgi?id=11413